India's adherence to principle of democracy - a comparison
Compares how the electoral democracy index of India has evolved since its independence in 1947. For context and comparison, the evolution of the same index over time of 2 countries each from Europe, Africa, North America, and South America are provided.

Expectedly, India comes somewhere in the middle, with the top positions being taken up by European and North American countries. The South American countries of Brazil and Chile have improved significantly over the last two decades while the African countries display huge variations (in line with the constant ostensible economic and political occurrences). Indonesia, since the huge improvement in 2000, has plateaued out whereas India has flatlined for a long time. 

This suggests India maybe in a bit of an equilibrium state and likely exhibits homeostasis in regard to its democracy. So while it is resistant on a macro scale to political upheavals like military coups and such, no improvements too seem to be happening. 

The one thing of note among all democracies worldwide is that there seems to be a downward swing, indicating deterioration, of adherence to democratic principles in this decade over the last few years. This warrants further study.